Avoid unsafe use of `char::from_u32_unchecked`.
The way this function is being used is directly violating its
safety clause. Just because the standard library does it internally,
it does not mean that it safe do it in non-standard crates.

That behavior is not documented and may change at any moment
(which includes crashing or producing a different result).

The current implementations of `encode_utf8` and `encode_utf16` have
been copied from Rust's libcore, changing the parameter type from `char`
to `u32`.